server/requirements.txt
# FastAPI Server Requirements
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# LangChain
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1

# Vector Store
faiss-cpu>=1.7.4

# OpenAI
openai>=1.3.0
